refactor: 소셜 로그인을 위한 커스텀 클래스 패키지명 변경
