+++ src/main/java/com/example/nyeon/auth/sociallogin/userservice/SocialOAuth2UserService.java
@@ -1,7 +1,7 @@
-package com.example.nyeon.auth.security.userservice;
+package com.example.nyeon.auth.sociallogin.userservice;
 
 import com.example.nyeon.auth.exception.BadRequestException;
-import com.example.nyeon.auth.security.principal.OAuth2UserPrincipal;
+import com.example.nyeon.auth.sociallogin.principal.OAuth2UserPrincipal;
 import com.example.nyeon.auth.user.UserRepository;
 import java.util.Map;
 import java.util.Optional;
